requirements.txt
# Core Consciousness Computing Dependencies
asyncio-mqtt>=0.11.1
websockets>=10.4
aiohttp>=3.8.4
httpx>=0.24.0

# Data Science & Machine Learning
numpy>=1.21.0
scipy>=1.7.0
scikit-learn>=1.0.0
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0

# Vector Matrix Orchestration
umap-learn>=0.5.3
hdbscan>=0.8.28
faiss-cpu>=1.7.3

# Web Framework & GUI
flask>=2.2.0
flask-cors>=4.0.0
dash>=2.6.0
plotly>=5.10.0

# Testing & Validation
pytest>=7.1.0
pytest-asyncio>=0.18.0
playwright>=1.28.0

# Logging & Monitoring
structlog>=22.1.0
rich>=12.5.0
tqdm>=4.64.0
psutil>=5.9.0

# =============================================================================
# OBSERVABILITY STACK - Ultimate 25-Domain Coverage
# =============================================================================

# OpenTelemetry Core
opentelemetry-api>=1.21.0
opentelemetry-sdk>=1.21.0
opentelemetry-semantic-conventions>=0.42b0

# OpenTelemetry Exporters
opentelemetry-exporter-otlp>=1.21.0
opentelemetry-exporter-otlp-proto-grpc>=1.21.0
opentelemetry-exporter-otlp-proto-http>=1.21.0
opentelemetry-exporter-prometheus>=0.42b0
opentelemetry-exporter-jaeger>=1.21.0

# OpenTelemetry Instrumentation
opentelemetry-instrumentation>=0.42b0
opentelemetry-instrumentation-fastapi>=0.42b0
opentelemetry-instrumentation-httpx>=0.42b0
opentelemetry-instrumentation-aiohttp-client>=0.42b0
opentelemetry-instrumentation-logging>=0.42b0
opentelemetry-instrumentation-requests>=0.42b0
opentelemetry-instrumentation-sqlalchemy>=0.42b0
opentelemetry-instrumentation-redis>=0.42b0
opentelemetry-instrumentation-grpc>=0.42b0

# OpenTelemetry Propagators
opentelemetry-propagator-b3>=1.21.0
opentelemetry-propagator-jaeger>=1.21.0

# Prometheus Client
prometheus-client>=0.19.0

# LLM Observability
langfuse>=2.0.0
# arize-phoenix>=3.0.0  # Optional: Self-hosted LLM observability

# ML Monitoring (Drift Detection)
evidently>=0.4.0

# Distributed Tracing Clients
# jaeger-client>=4.8.0  # Deprecated, use OTLP
py-zipkin>=1.0.0

# Chaos Engineering
# steadybit-agent  # Installed separately

# Incident Management
# pagerduty>=2.0.0  # Optional: PagerDuty integration

# CloudEvents
cloudevents>=1.10.0

# Data Lake / Analytics
pyarrow>=14.0.0
# apache-iceberg>=0.5.0  # Optional: Data lake support

# FastAPI/Starlette (if not already present)
fastapi>=0.104.0
starlette>=0.27.0
uvicorn>=0.24.0

# Security & Cryptography
cryptography>=38.0.0
bcrypt>=4.0.0
pyjwt>=2.6.0

# Configuration & Serialization
pyyaml>=6.0
toml>=0.10.2
orjson>=3.8.0
jsonschema>=4.17.0
pydantic>=2.0.0

# Development Tools
black>=22.6.0
isort>=5.10.0
flake8>=4.0.0
mypy>=0.971
ruff>=0.1.0
safety>=2.3.0

# Optional: Quantum Computing (future)
# qiskit>=0.39.0

# Optional: Advanced ML (future)
# torch>=1.12.0
# transformers>=4.21.0
# accelerate>=0.15.0

# Consciousness-Specific Libraries
# (Custom implementations in consciousness_suite/)
# Recursive Enlightenment Engine
# Vector Matrix Orchestrator
# Ultra-Critic Swarm
# ABYSSAL Template System
